backend/requirements.txt
pydantic==2.12.5
google-genai==1.56.0
pillow==12.0.0
uvicorn==0.40.0
fastapi==0.128.0
dotenv==0.9.9
python-multipart==0.0.21
SQLAlchemy==1.4.39
alembic==1.8.1
fastapi-cache2==0.2.2
minio==7.2.20
asyncpg==0.31.0
greenlet==3.3.0
psycopg2==2.9.11
jinja2==3.0.3
authlib==1.3.0
httpx==0.27.0
PyJWT==2.10.1
itsdangerous==2.2.0